improved on numeric and vector control

+++ src/ModelUpdateView/NumericControlModel.g.fs
@@ -39,7 +39,7 @@ module Mutable =
         [<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
         module Lens =
             let value =
-                { new Lens<NumericControlNs.NumericModel, Microsoft.FSharp.Core.int>() with
+                { new Lens<NumericControlNs.NumericModel, Microsoft.FSharp.Core.float>() with
                     override x.Get(r) = r.value
                     override x.Set(r,v) = { r with value = v }
                     override x.Update(r,f) = { r with value = f r.value }
